+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## CHANGE LOG
 
 - v0.1.5
-  - add JsonCodec support
+  - add JsonCodec/ProtoCodec support